refactor: update to @angular 9.0.0 & cypress 4.0.0 (#30)
* refactor(ui/frontend): import deeply to specific material component

* build(dep): update `@angular/cli` to `v8.3.25`

* build(dep): update `@nrwl/*` to `v8.12.2`

* build(dep): update `@angular/*` to `v9`

* build(dep): update `@angular/material` to `v9`

* build(dep): update `@storybook`, `@typescript-eslint`

* build(dep): update `cypress` to `v4.0.0`

* build(dep): remove `handlebars`

* build(dep): update `@babylonjs`
