+++ apps/frontend/src/app/options-panel/options-panel.module.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { MatButtonModule, MatIconModule } from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
 import { EffectsModule } from '@ngrx/effects';
 import { UiColorDialogModule } from '@talus/ui';
 import { OptionsPanelComponent } from './options-panel.component';
