+++ apps/frontend/src/main.ts
@@ -1,6 +1,5 @@
 import { enableProdMode, ViewEncapsulation } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import 'hammerjs';
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
