+++ apps/frontend/src/test-setup.ts
@@ -1,2 +1,1 @@
-import 'hammerjs';
 import 'jest-preset-angular';
