+++ libs/ui/src/lib/color-dialog/color-dialog.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { MatDialog, MatDialogRef } from '@angular/material';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import {
   UiColorDialogColor,
   UiColorDialogComponent,
