+++ libs/ui/src/lib/toolbar/toolbar.module.ts
@@ -1,6 +1,8 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { MatButtonToggleModule, MatIconModule, MatTooltipModule } from '@angular/material';
+import { MatButtonToggleModule } from '@angular/material/button-toggle';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTooltipModule } from '@angular/material/tooltip';
 import { UiToolbarComponent } from './toolbar.component';
 
 @NgModule({
