+++ package.json
@@ -22,6 +22,7 @@
     "lint:styles": "stylelint \"{apps,libs}/**/*.{css,scss}\"",
     "ng": "ng",
     "nx": "nx",
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
     "prettier:check": "yarn prettier --check --config ./.prettierrc --ignore-path ./.prettierignore '{*,\\.circleci,apps,libs}/**/*.{ts,scss,json,html,yml}'",
     "start": "ng serve",
     "test": "ng test",
@@ -32,56 +33,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/cdk": "8.2.3",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/material": "8.2.3",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
-    "@babylonjs/core": "4.1.0-beta.21",
-    "@babylonjs/materials": "4.1.0-beta.21",
+    "@angular/animations": "9.0.0",
+    "@angular/cdk": "9.0.0",
+    "@angular/common": "9.0.0",
+    "@angular/compiler": "9.0.0",
+    "@angular/core": "9.0.0",
+    "@angular/forms": "9.0.0",
+    "@angular/material": "9.0.0",
+    "@angular/platform-browser": "9.0.0",
+    "@angular/platform-browser-dynamic": "9.0.0",
+    "@angular/router": "9.0.0",
+    "@babylonjs/core": "4.1.0-beta.23",
+    "@babylonjs/materials": "4.1.0-beta.23",
     "@ngrx/effects": "8.6.0",
     "@ngrx/store": "8.6.0",
-    "@nrwl/angular": "8.12.1",
-    "hammerjs": "2.0.8",
+    "@nrwl/angular": "8.12.2",
     "mnemonist": "0.32.0",
     "rxjs": "6.5.4",
+    "tslib": "1.10.0",
     "zone.js": "0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.803.24",
-    "@angular/cli": "8.3.24",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "0.900.1",
+    "@angular/cli": "9.0.1",
+    "@angular/compiler-cli": "9.0.0",
+    "@angular/language-service": "9.0.0",
     "@babel/core": "7.8.4",
-    "@babylonjs/inspector": "4.1.0-beta.21",
-    "@nrwl/cypress": "8.12.1",
-    "@nrwl/eslint-plugin-nx": "8.12.1",
-    "@nrwl/jest": "8.12.1",
-    "@nrwl/node": "8.12.1",
-    "@nrwl/storybook": "8.12.1",
-    "@nrwl/workspace": "8.12.1",
-    "@storybook/addon-knobs": "5.3.10",
-    "@storybook/angular": "5.3.10",
+    "@babylonjs/inspector": "4.1.0-beta.23",
+    "@nrwl/cypress": "8.12.2",
+    "@nrwl/eslint-plugin-nx": "8.12.2",
+    "@nrwl/jest": "8.12.2",
+    "@nrwl/node": "8.12.2",
+    "@nrwl/storybook": "8.12.2",
+    "@nrwl/workspace": "8.12.2",
+    "@storybook/addon-knobs": "5.3.12",
+    "@storybook/angular": "5.3.12",
     "@types/benchmark": "1.0.31",
     "@types/jest": "24.9.0",
     "@types/node": "13.7.0",
-    "@typescript-eslint/eslint-plugin": "2.18.0",
-    "@typescript-eslint/parser": "2.18.0",
+    "@typescript-eslint/eslint-plugin": "2.19.0",
+    "@typescript-eslint/parser": "2.19.0",
     "babel-loader": "8.0.6",
-    "babylonjs": "4.1.0-beta.21",
+    "babylonjs": "4.1.0-beta.23",
     "benchmark": "2.1.4",
     "codelyzer": "5.2.1",
-    "cypress": "3.8.3",
+    "cypress": "4.0.0",
     "dotenv": "8.2.0",
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.10.0",
     "gh-pages": "2.2.0",
-    "handlebars": "4.7.2",
     "jasmine-marbles": "0.6.0",
     "jest": "24.9.0",
     "jest-junit": "10.0.0",
@@ -96,6 +96,6 @@
     "ts-jest": "24.3.0",
     "ts-node": "8.6.2",
     "tslint": "5.20.1",
-    "typescript": "3.5.3"
+    "typescript": "3.7.5"
   }
 }
